CSC4350/project1/requirements.txt
Flask
requests
python-dotenv